src/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2 
gunicorn==23.0.0
openmeteo-requests==1.3.0
psycopg[binary,pool]==3.2.6
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
sqlparse==0.5.3